# lib/api-spec/openapi.yaml
openapi: "3.1.0"
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: "1.0.0"
  description: Sakani Dz - Real Estate & Tourism Platform API
servers:
  - url: /api
    description: Base API path
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AuthResponse:
      type: object
      required: [accessToken, refreshToken, role, userId]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        role:
          type: string
          enum: [client, vendor, admin]
        userId:
          type: integer
        name:
          type: ["string", "null"]

    # â”€â”€ Entities â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€
...
    ClientRegistrationInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
        password:
          type: string
          minLength: 6
        phone:
          type: string
        preferredWilaya:
          type: string
    ClientLoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    VendorLoginInput:
      type: object
      required: [code, password]
      properties:
        code:
          type: string
        password:
          type: string
    AdminLoginInput:
      type: object
      required: [password]
      properties:
        password:
          type: string
    TokenRefreshInput:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    ForgotPasswordInput:
      type: object
      required: [code]
      properties:
        code:
          type: string
    AnonymousLoginInput:
      type: object
      required: [deviceId]
      properties:
        deviceId:
          type: string
...
paths:
...
  /auth/client/register:
    post:
      operationId: registerClient
      tags: [auth]
      summary: Register a new client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientRegistrationInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/client/login:
    post:
      operationId: loginClient
      tags: [auth]
      summary: Client login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientLoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/vendor/login:
    post:
      operationId: loginVendor
      tags: [auth]
      summary: Vendor login with code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorLoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/admin/login:
    post:
      operationId: loginAdmin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenRefreshInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/vendor/forgot-password:
    post:
      operationId: vendorForgotPassword
      tags: [auth]
      summary: Vendor forgot password request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordInput"
      responses:
        "200":
          description: Request submitted
          content:
            application/json:
              schema:
                type: object
                required: [message]
                properties:
                  message:
                    type: string
  /auth/client/anonymous:
    post:
      operationId: loginAnonymousClient
      tags: [auth]
      summary: Get or create an anonymous client session tied to a device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnonymousLoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"

  # â”€â”€ Client profile â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€
...